rust: fix clippy lints for clippy::unnecessary_cast

+++ rust/src/telnet/telnet.rs
@@ -176,7 +176,7 @@ impl TelnetState {
                     flow,
                     stream_slice,
                     start,
-                    -1 as i64,
+                    -1_i64,
                     TelnetFrameType::Pdu as u8,
                 );
             }
@@ -187,7 +187,7 @@ impl TelnetState {
                             flow,
                             stream_slice,
                             start,
-                            -1 as i64,
+                            -1_i64,
                             TelnetFrameType::Ctl as u8,
                         )
                     } else {
@@ -195,7 +195,7 @@ impl TelnetState {
                             flow,
                             stream_slice,
                             start,
-                            -1 as i64,
+                            -1_i64,
                             TelnetFrameType::Data as u8,
                         )
                     // app-layer-frame-documentation tag end: parse_request
@@ -280,14 +280,14 @@ impl TelnetState {
         let mut start = input;
         while !start.is_empty() {
             if self.response_frame.is_none() {
-                self.response_frame = Frame::new(flow, stream_slice, start, -1 as i64, TelnetFrameType::Pdu as u8);
+                self.response_frame = Frame::new(flow, stream_slice, start, -1_i64, TelnetFrameType::Pdu as u8);
             }
             if self.response_specific_frame.is_none() {
                 if let Ok((_, is_ctl)) = parser::peek_message_is_ctl(start) {
                     self.response_specific_frame = if is_ctl {
-                        Frame::new(flow, stream_slice, start, -1 as i64, TelnetFrameType::Ctl as u8)
+                        Frame::new(flow, stream_slice, start, -1_i64, TelnetFrameType::Ctl as u8)
                     } else {
-                        Frame::new(flow, stream_slice, start, -1 as i64, TelnetFrameType::Data as u8)
+                        Frame::new(flow, stream_slice, start, -1_i64, TelnetFrameType::Data as u8)
                     };
                 }
             }
